+++ api/src/service/domain/workflow/workflowitem_create.spec.ts
@@ -5,6 +5,8 @@ import * as Result from "../../../result";
 import { PreconditionError } from "../errors/precondition_error";
 import { ServiceUser } from "../organization/service_user";
 import { Subproject } from "./subproject";
+import { UserRecord } from "../organization/user_record";
+import { Permissions } from "../permissions";
 import * as WorkflowitemCreate from "./workflowitem_create";
 
 const ctx: Ctx = { requestId: "", source: "test" };
@@ -29,6 +31,32 @@ const baseSubproject: Subproject = {
   additionalData: {},
 };
 
+const normalUser: UserRecord = {
+  id: "test",
+  createdAt: new Date().toISOString(), // ISO timestamp
+  displayName: "testuser",
+  organization: "Test Org",
+  passwordHash: "abc",
+  address: "testaddress",
+  encryptedPrivKey: "abcd",
+  permissions: {["user.authenticate"]: ["test"]},
+  log: [],
+  additionalData: {}
+}
+
+const disabledUser: UserRecord = {
+  id: "test",
+  createdAt: new Date().toISOString(), // ISO timestamp
+  displayName: "testuser",
+  organization: "Test Org",
+  passwordHash: "abc",
+  address: "testaddress",
+  encryptedPrivKey: "abcd",
+  permissions: {}, // no auth permissions
+  log: [],
+  additionalData: {}
+}
+
 describe("Create workflowitem", () => {
   it("Root cannot create a workflow item", async () => {
     const data: WorkflowitemCreate.RequestData = {
@@ -41,6 +69,8 @@ describe("Create workflowitem", () => {
 
     const result = await WorkflowitemCreate.createWorkflowitem(ctx, root, data, {
       workflowitemExists: async (_projectId, _subprojectId, _workflowitemId) => false,
+      userExists: async (_userId) => true,
+      getUser: async (_userId) => normalUser,
       getSubproject: async () => baseSubproject,
       applyWorkflowitemType: () => [],
       uploadDocumentToStorageService: () => Promise.resolve([]),
@@ -62,6 +92,54 @@ describe("Create workflowitem", () => {
 
     const result = await WorkflowitemCreate.createWorkflowitem(ctx, alice, data, {
       workflowitemExists: async (_projectId, _subprojectId, _workflowitemId) => false,
+      userExists: async (_userId) => true,
+      getUser: async (_userId) => normalUser,
+      getSubproject: async () => baseSubproject,
+      applyWorkflowitemType: () => [],
+      uploadDocumentToStorageService: () => Promise.resolve([]),
+      getAllDocumentReferences: async () => [],
+    });
+
+    assert.isTrue(Result.isErr(result));
+    assert.instanceOf(Result.unwrapErr(result), PreconditionError);
+  });
+
+  it("Cannot create a workflow item if the assigned user does not exist!", async () => {
+    const data: WorkflowitemCreate.RequestData = {
+      projectId: "test",
+      subprojectId: "dummy-subproject",
+      displayName: "test",
+      amountType: "N/A",
+      workflowitemType: "general",
+    };
+
+    const result = await WorkflowitemCreate.createWorkflowitem(ctx, alice, data, {
+      workflowitemExists: async (_projectId, _subprojectId, _workflowitemId) => false,
+      userExists: async (_userId) => false,
+      getUser: async (_userId) => normalUser,
+      getSubproject: async () => baseSubproject,
+      applyWorkflowitemType: () => [],
+      uploadDocumentToStorageService: () => Promise.resolve([]),
+      getAllDocumentReferences: async () => [],
+    });
+
+    assert.isTrue(Result.isErr(result));
+    assert.instanceOf(Result.unwrapErr(result), PreconditionError);
+  });
+
+  it("Cannot create a workflow item if the assigned user is disabled!", async () => {
+    const data: WorkflowitemCreate.RequestData = {
+      projectId: "test",
+      subprojectId: "dummy-subproject",
+      displayName: "test",
+      amountType: "N/A",
+      workflowitemType: "general",
+    };
+
+    const result = await WorkflowitemCreate.createWorkflowitem(ctx, alice, data, {
+      workflowitemExists: async (_projectId, _subprojectId, _workflowitemId) => false,
+      userExists: async (_userId) => false,
+      getUser: async (_userId) => disabledUser,
       getSubproject: async () => baseSubproject,
       applyWorkflowitemType: () => [],
       uploadDocumentToStorageService: () => Promise.resolve([]),